Handler tests now pass

+++ src/Forms/test/Xamarin.Forms.Core.UnitTests/BoxViewUnitTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Graphics;
 using System.Linq;
 using NUnit.Framework;
 
@@ -13,14 +14,14 @@ namespace Xamarin.Forms.Core.UnitTests
 		{
 			var box = new BoxView
 			{
-				Color = new Color(0.2, 0.3, 0.4),
+				Color = new Color(0.2f, 0.3f, 0.4f),
 				WidthRequest = 20,
 				HeightRequest = 30,
 				IsPlatformEnabled = true,
 			};
 
-			Assert.AreEqual(new Color(0.2, 0.3, 0.4), box.Color);
-			var request = box.GetSizeRequest(double.PositiveInfinity, double.PositiveInfinity).Request;
+			Assert.AreEqual(new Color(0.2f, 0.3f, 0.4f), box.Color);
+			var request = box.GetSizeRequest(float.PositiveInfinity, float.PositiveInfinity).Request;
 			Assert.AreEqual(20, request.Width);
 			Assert.AreEqual(30, request.Height);
 		}
@@ -33,7 +34,7 @@ namespace Xamarin.Forms.Core.UnitTests
 				IsPlatformEnabled = true,
 			};
 
-			var request = box.GetSizeRequest(double.PositiveInfinity, double.PositiveInfinity).Request;
+			var request = box.GetSizeRequest(float.PositiveInfinity, float.PositiveInfinity).Request;
 			Assert.AreEqual(40, request.Width);
 			Assert.AreEqual(40, request.Height);
 		}
