+++ src/Platform.Handlers/tests/Xamarin.Platform.Handlers.UnitTests/Layouts/ConstraintTests.cs
@@ -11,7 +11,7 @@ namespace Xamarin.Platform.Handlers.UnitTests.Layouts
 		[TestCase(100, 200, 100)]
 		[TestCase(100, 100, 100)]
 		[TestCase(100, -1, 100)]
-		public void ExternalWinsOverDesired(double externalConstraint, double desiredLength, double expected) 
+		public void ExternalWinsOverDesired(float externalConstraint, float desiredLength, float expected) 
 		{
 			var resolution = LayoutManager.ResolveConstraints(externalConstraint, desiredLength);
 			Assert.That(resolution, Is.EqualTo(expected));
@@ -20,7 +20,7 @@ namespace Xamarin.Platform.Handlers.UnitTests.Layouts
 		[Test(Description = "When resolving constraints, external constraints take precedence")]
 		[TestCase(100, 200, 130, 100)]
 		[TestCase(100, -1, 130, 100)]
-		public void ExternalWinsOverDesired(double externalConstraint, double desiredLength, double measured, double expected)
+		public void ExternalWinsOverDesired(float externalConstraint, float desiredLength, float measured, float expected)
 		{
 			var resolution = LayoutManager.ResolveConstraints(externalConstraint, desiredLength, measured);
 			Assert.That(resolution, Is.EqualTo(expected));
@@ -29,14 +29,14 @@ namespace Xamarin.Platform.Handlers.UnitTests.Layouts
 		[Test(Description = "If external and request constraints don't apply, constrain to measured value")]
 		public void MeasuredWinsIfNothingElseApplies()
 		{
-			var resolution = LayoutManager.ResolveConstraints(double.PositiveInfinity, -1, 245);
+			var resolution = LayoutManager.ResolveConstraints(float.PositiveInfinity, -1, 245);
 			Assert.That(resolution, Is.EqualTo(245));
 		}
 
 		[Test(Description = "If external constraints don't apply, constraing to requested value")]
 		public void RequestedTakesPrecedenceOverMeasured()
 		{
-			var resolution = LayoutManager.ResolveConstraints(double.PositiveInfinity, 90, 245);
+			var resolution = LayoutManager.ResolveConstraints(float.PositiveInfinity, 90, 245);
 			Assert.That(resolution, Is.EqualTo(90));
 		}
 	}
